+++ demo/src/main/java/com/example/demo/repositories/ProjetoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.models.Projeto;
 
-public interface ProjetoRepository extends JpaRepository<Projeto, Integer>{
+public interface ProjetoRepository extends JpaRepository<Projeto, Long>{
     
 }
