+++ demo/src/main/java/com/example/demo/repositories/PropostaRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.models.Proposta;
 
-public interface PropostaRepository extends JpaRepository<Proposta, Integer> {
+public interface PropostaRepository extends JpaRepository<Proposta, Long> {
     
 }
